Split frame representation from solid representation
